+++ parser/src/ast/array_len.rs
@@ -1,6 +1,7 @@
 use super::{LitPosInt, NamedRef};
 use crate::error::{
-    ConstIntNotFound, ExpectedConstIntFoundService, ExpectedConstIntFoundType, InvalidArrayLen,
+    ConstIntNotFound, ExpectedConstIntFoundService, ExpectedConstIntFoundString,
+    ExpectedConstIntFoundType, InvalidArrayLen,
 };
 use crate::grammar::Rule;
 use crate::validate::Validate;
@@ -63,6 +64,7 @@ impl ArrayLenValue {
             Self::Ref(ty) => {
                 ConstIntNotFound::validate(ty, validate);
                 ExpectedConstIntFoundService::validate(ty, validate);
+                ExpectedConstIntFoundString::validate(ty, validate);
                 ExpectedConstIntFoundType::validate(ty, validate);
 
                 ty.validate(validate);