+++ parser/src/test/ui_tests.rs
@@ -1,6 +1,7 @@
 ui_test!(const_int_not_found);
 ui_test!(error_empty_enum);
 ui_test!(expected_const_int_found_service);
+ui_test!(expected_const_int_found_string);
 ui_test!(expected_const_int_found_type);
 ui_test!(expected_type_found_const);
 ui_test!(expected_type_found_service);